Rename files from cr2image_int to cr2header_int
